Working version of i18n
I followed the steps on this blog post[1] to setup i18n. It involved
making the decision to add a locale code as a URL segment. I may change
this later but it seems sound for now.

I also went through and wrapped all app text in "gettext" calls but have
yet to generate the po files. That will be next.

[1] http://code.parent.co/practical-i18n-with-phoenix-and-elixir/