Add functionality to pull exif metadata from photos
This involves adding ExExif to pull the exif & gps data from the
uploaded photo and Timex to parse the datetime the photo was taken into
Elixir native DateTime.

I don't think the code I added is in the correct place, but I'm not
totally sure where is should go...so I'm leaving it in the controller
for now.