Large refactor of locale plug
I was trying to update the auth tests and realized that I didn't have a
strong grasp on what the locale plug did (I copied it from a blog post).
The code was a little hard to understand so I wanted to start renaming
stuff. Then I remembered that there were no tests for this plug :(
So then I:
- wrote tests for the locale plug
- refactored it to make more sense (to me anyway)
  - lots of renaming
  - removed Ruby-esque calls to Blank protocol
  - moved it into the web/plugs folder

I also realized that locale is a misnomer in this case - I believe we
are only working with languages here, so I want to refresh myself on
this a bit and come back for another renaming refactor :)