Update auth test to work with our localization
This may only be a one line change, but it took me all morning to get
here. Originally this test was broken and I knew it was because of the
localization routing stuff we'd added, but I didn't have a strong
enough grasp on what the localization routing was doing...so I wrote
tests and refactored the localization plug. Doing that made it so that
when I came back to this test, the change I needed to make was obvious.